Added tests for convertTX
