+++ packages/bitcore-node/src/providers/adapter/bch/BchAdapter.ts
