+++ packages/bitcore-node/src/providers/adapter/btc/BtcAdapter.ts
